Refactor the event conversion code
