+++ crates/core/c8y_api/Cargo.toml
@@ -15,6 +15,7 @@ clock = { path = "../../common/clock" }
 csv = "1.1"
 download = { path = "../../common/download" }
 futures = "0.3"
+json-writer = { path = "../../common/json_writer" }
 mockall = "0.11"
 mqtt_channel = { path = "../../common/mqtt_channel" }
 reqwest = { version = "0.11", default-features = false, features = ["json", "rustls-tls"] }
