+++ crates/core/c8y_smartrest/Cargo.toml
@@ -10,7 +10,9 @@ agent_interface = { path = "../agent_interface" }
 csv = "1.1"
 download = { path = "../../common/download" }
 mqtt_channel = { path = "../../common/mqtt_channel" }
+json-writer = { path = "../../common/json_writer" }
 reqwest = { version = "0.11", default-features = false, features = ["json", "rustls-tls"] }
+serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 tedge_config = { path = "../../common/tedge_config" }
 thin_edge_json = { path = "../thin_edge_json" }
