+++ crates/core/tedge_mapper/Cargo.toml
@@ -38,7 +38,6 @@ download = { path = "../../common/download" }
 flockfile = { path = "../../common/flockfile" }
 futures = "0.3"
 inotify = "0.10.0"
-json-writer = { path = "../../common/json_writer" }
 logged_command = { path = "../../common/logged_command" }
 mockall = "0.11"
 mqtt_channel = { path = "../../common/mqtt_channel" }
